Random projection for dimensionality reduction.

Generate a random projection matrix for dimensionality reduction, and optionally transform input data to a projection in a lower dimension space using the generated random matrix.

cuda_ml_rand_proj(
  x,
  n_components = NULL,
  eps = 0.1,
  gaussian_method = TRUE,
  density = NULL,
  transform_input = TRUE,
  seed = 0L
)

Arguments

x
The input matrix or dataframe. Each data point should be a row and should consist of numeric values only.

n_components
Dimensionality of the target projection space. If NULL, then the parameter is deducted using the Johnson-Lindenstrauss lemma, taking into consideration the number of samples and the eps parameter. Default: NULL.

eps
Error tolerance during projection. Default: 0.1.

gaussian_method
If TRUE, then use the Gaussian random projection method. Otherwise, use the sparse random projection method. See https://en.wikipedia.org/wiki/Random_projection for details. Default: TRUE.

density
Ratio of non-zero component in the random projection matrix. If NULL, then the value is set to the minimum density as recommended by Ping Li et al.: 1 / sqrt(n_features). Default: NULL.

transform_input
Whether to project input data onto a lower dimension space using the random matrix. Default: TRUE.

seed
Seed for the pseudorandom number generator. Default: 0L.

Value

A context object containing GPU pointer to a random matrix that can be used as input to the cuda_ml_transform() function. If transform_input is set to TRUE, then the context object will also contain a "transformed_data" attribute containing the lower dimensional projection of the input data.

Examples

library(cuda.ml)
library(mlbench)

data(Vehicle)
vehicle_data <- Vehicle[order(Vehicle$Class), which(names(Vehicle) != "Class")]

model <- cuda_ml_rand_proj(vehicle_data, n_components = 4)

set.seed(0L)
print(kmeans(model$transformed_data, centers = 4, iter.max = 1000))

#> K-means clustering with 4 clusters of sizes 843, 1, 1, 1
#> 
#> Cluster means:
#>         [,1]       [,2]       [,3]       [,4]       [,5]       [,6]       [,7]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>         [,8]       [,9]      [,10]      [,11]      [,12]      [,13]      [,14]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>        [,15]      [,16]      [,17]      [,18]      [,19]      [,20]      [,21]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>        [,22]      [,23]      [,24]      [,25]      [,26]      [,27]      [,28]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>        [,29]      [,30]      [,31]      [,32]      [,33]      [,34]      [,35]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>        [,36]      [,37]      [,38]      [,39]      [,40]      [,41]      [,42]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>        [,43]      [,44]      [,45]      [,46]      [,47]      [,48]      [,49]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>        [,50]      [,51]      [,52]      [,53]      [,54]      [,55]      [,56]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>        [,57]      [,58]      [,59]      [,60]      [,61]      [,62]      [,63]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>        [,64]      [,65]      [,66]      [,67]      [,68]      [,69]      [,70]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>        [,71]      [,72]      [,73]      [,74]      [,75]      [,76]      [,77]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>        [,78]      [,79]      [,80]      [,81]      [,82]      [,83]      [,84]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>        [,85]      [,86]      [,87]      [,88]      [,89]      [,90]      [,91]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>        [,92]      [,93]      [,94]      [,95]      [,96]      [,97]      [,98]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>        [,99]     [,100]     [,101]     [,102]     [,103]     [,104]     [,105]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,106]     [,107]     [,108]     [,109]     [,110]     [,111]     [,112]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,113]     [,114]     [,115]     [,116]     [,117]     [,118]     [,119]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,120]     [,121]     [,122]     [,123]     [,124]     [,125]     [,126]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,127]     [,128] [,129]     [,130]     [,131]     [,132]     [,133]
#> 1 0.00118624 0.00118624      0 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000      0 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000      0 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000      1 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,134]     [,135]     [,136]     [,137]     [,138]     [,139]     [,140]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,141]     [,142]     [,143]     [,144]     [,145]     [,146]     [,147]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,148]     [,149]     [,150]     [,151]     [,152]     [,153]     [,154]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,155]     [,156]     [,157]     [,158]     [,159]     [,160]     [,161]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,162]     [,163]     [,164]     [,165]     [,166]     [,167]     [,168]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,169]     [,170]     [,171]     [,172]     [,173]     [,174]     [,175]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,176]     [,177]     [,178]     [,179]     [,180]     [,181]     [,182]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,183]     [,184]     [,185]     [,186]     [,187]     [,188]     [,189]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,190]     [,191]     [,192]     [,193]     [,194]     [,195]     [,196]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,197]     [,198]     [,199]     [,200]     [,201]     [,202]     [,203]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,204]     [,205]     [,206]     [,207]     [,208]     [,209]     [,210]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,211]     [,212]     [,213]     [,214]     [,215]     [,216]     [,217]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,218]     [,219]     [,220]     [,221]     [,222]     [,223]     [,224]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,225]     [,226]     [,227]     [,228]     [,229]     [,230]     [,231]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,232]     [,233]     [,234]     [,235]     [,236]     [,237]     [,238]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,239]     [,240]     [,241]     [,242]     [,243]     [,244]     [,245]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,246]     [,247]     [,248]     [,249]     [,250]     [,251]     [,252]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,253]     [,254]     [,255]     [,256]     [,257]     [,258]     [,259]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,260]     [,261]     [,262]     [,263]     [,264]     [,265]     [,266]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,267]     [,268]     [,269]     [,270]     [,271]     [,272]     [,273]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,274]     [,275]     [,276]     [,277]     [,278]     [,279]     [,280]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,281]     [,282]     [,283]     [,284]     [,285]     [,286]     [,287]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,288]     [,289]     [,290]     [,291]     [,292]     [,293]     [,294]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,295]     [,296]     [,297]     [,298]     [,299]     [,300]     [,301]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,302]     [,303]     [,304]     [,305]     [,306]     [,307]     [,308]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,309]     [,310]     [,311]     [,312]     [,313]     [,314]     [,315]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,316]     [,317]     [,318]     [,319]     [,320]     [,321]     [,322]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,323]     [,324]     [,325]     [,326]     [,327]     [,328]     [,329]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,330]     [,331]     [,332]     [,333]     [,334]     [,335]     [,336]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,337]     [,338]     [,339]     [,340]     [,341]     [,342]     [,343]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,344]     [,345]     [,346]     [,347]     [,348]     [,349]     [,350]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,351]     [,352]     [,353]     [,354]     [,355]     [,356]     [,357]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,358]     [,359]     [,360]     [,361]     [,362]     [,363]     [,364]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,365]     [,366]     [,367]     [,368]     [,369]     [,370]     [,371]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,372]     [,373]     [,374]     [,375]     [,376]     [,377]     [,378]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,379]     [,380]     [,381]     [,382]     [,383]     [,384]     [,385]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,386]     [,387]     [,388]     [,389]     [,390]     [,391]     [,392]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,393]     [,394]     [,395]     [,396]     [,397]     [,398]     [,399]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,400]     [,401]     [,402]     [,403]     [,404]     [,405]     [,406]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,407]     [,408]     [,409]     [,410]     [,411]     [,412]     [,413]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,414]     [,415]     [,416]     [,417]     [,418]     [,419]     [,420]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,421]     [,422]     [,423]     [,424]     [,425]     [,426]     [,427]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,428]     [,429]     [,430]     [,431]     [,432]     [,433]     [,434]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,435]     [,436]     [,437]     [,438]     [,439]     [,440]     [,441]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,442]     [,443]     [,444]     [,445]     [,446]     [,447]     [,448]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,449]     [,450]     [,451]     [,452]     [,453]     [,454]     [,455]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,456]     [,457]     [,458]     [,459]     [,460]     [,461]     [,462]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,463]     [,464]     [,465]     [,466]     [,467]     [,468]     [,469]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,470]     [,471]     [,472]     [,473]     [,474]     [,475]     [,476]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,477]     [,478]     [,479]     [,480]     [,481]     [,482]     [,483]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,484]     [,485]     [,486]     [,487]     [,488]     [,489]     [,490]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,491]     [,492]     [,493]     [,494]     [,495]     [,496]     [,497]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,498]     [,499]     [,500]     [,501]     [,502]     [,503]     [,504]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,505]     [,506]     [,507]     [,508]     [,509]     [,510]     [,511]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,512]     [,513]     [,514]     [,515]     [,516]     [,517]     [,518]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,519]     [,520]     [,521]     [,522]     [,523]     [,524]     [,525]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,526]     [,527]     [,528]     [,529]     [,530]     [,531]     [,532]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,533]     [,534]     [,535]     [,536]     [,537]     [,538]     [,539]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,540]     [,541]     [,542]     [,543]     [,544]     [,545]     [,546]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,547]     [,548]     [,549]     [,550]     [,551]     [,552]     [,553]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,554]     [,555]     [,556]     [,557]     [,558]     [,559]     [,560]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,561]     [,562]     [,563]     [,564]     [,565]     [,566]     [,567]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,568]     [,569]     [,570]     [,571]     [,572]     [,573]     [,574]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,575]     [,576]     [,577]     [,578]     [,579]     [,580]     [,581]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,582]     [,583]     [,584]     [,585]     [,586]     [,587]     [,588]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,589]     [,590]     [,591]     [,592]     [,593]     [,594]     [,595]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,596]     [,597]     [,598]     [,599]     [,600]     [,601]     [,602]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,603]     [,604]     [,605]     [,606]     [,607]     [,608]     [,609]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,610]     [,611]     [,612]     [,613]     [,614]     [,615]     [,616]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,617]     [,618]     [,619]     [,620]     [,621]     [,622]     [,623]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,624]     [,625]     [,626]     [,627]     [,628]     [,629]     [,630]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,631]     [,632]     [,633]     [,634]     [,635]     [,636]     [,637]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,638]     [,639]     [,640]     [,641]     [,642]     [,643]     [,644]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,645]     [,646]     [,647]     [,648]     [,649]     [,650]     [,651]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,652]     [,653]     [,654]     [,655]     [,656]     [,657]     [,658]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,659]     [,660]     [,661]     [,662]     [,663]     [,664]     [,665]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,666]     [,667]     [,668]     [,669]     [,670]     [,671]     [,672]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,673]     [,674]     [,675]     [,676]     [,677]     [,678] [,679]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624      0
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000      0
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000      1
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000      0
#>       [,680]     [,681]     [,682]     [,683]     [,684]     [,685]     [,686]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,687]     [,688]     [,689]     [,690]     [,691]     [,692]     [,693]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,694]     [,695]     [,696]     [,697]     [,698]     [,699]     [,700]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,701]     [,702]     [,703]     [,704]     [,705]     [,706]     [,707]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,708]     [,709]     [,710]     [,711]     [,712]     [,713]     [,714]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,715]     [,716]     [,717]     [,718]     [,719]     [,720]     [,721]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,722]     [,723]     [,724]     [,725]     [,726]     [,727]     [,728]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,729]     [,730]     [,731]     [,732]     [,733]     [,734]     [,735]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,736]     [,737]     [,738]     [,739]     [,740]     [,741]     [,742]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,743]     [,744]     [,745]     [,746]     [,747]     [,748]     [,749]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,750]     [,751]     [,752]     [,753]     [,754]     [,755]     [,756]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,757]     [,758]     [,759]     [,760]     [,761]     [,762]     [,763]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,764]     [,765]     [,766]     [,767]     [,768]     [,769]     [,770]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,771]     [,772]     [,773]     [,774]     [,775]     [,776]     [,777]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,778]     [,779]     [,780]     [,781]     [,782]     [,783]     [,784]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,785]     [,786]     [,787]     [,788]     [,789]     [,790]     [,791]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,792]     [,793]     [,794]     [,795]     [,796]     [,797]     [,798]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,799]     [,800]     [,801]     [,802]     [,803]     [,804]     [,805]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,806]     [,807]     [,808]     [,809]     [,810]     [,811]     [,812]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,813]     [,814]     [,815]     [,816]     [,817]     [,818]     [,819]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,820]     [,821]     [,822]     [,823]     [,824]     [,825]     [,826]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,827]     [,828]     [,829]     [,830]     [,831]     [,832]     [,833]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,834]     [,835] [,836]     [,837]     [,838]     [,839]     [,840]
#> 1 0.00118624 0.00118624      0 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000      1 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000      0 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000      0 0.00000000 0.00000000 0.00000000 0.00000000
#>       [,841]     [,842]     [,843]     [,844]     [,845]     [,846]
#> 1 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624 0.00118624
#> 2 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 
#> Clustering vector:
#>   [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [38] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [75] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [112] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 4 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [149] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [186] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [223] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [260] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [297] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [334] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [371] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [408] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [445] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [482] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [519] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [556] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [593] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [630] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [667] 1 1 1 1 1 1 1 1 1 1 1 1 3 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [704] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [741] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [778] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [815] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 2 1 1 1 1 1 1 1 1 1 1
#> 
#> Within cluster sum of squares by cluster:
#> [1] 842   0   0   0
#>  (between_SS / total_SS =   0.4 %)
#> 
#> Available components:
#> 
#> [1] "cluster"      "centers"      "totss"        "withinss"     "tot.withinss"
#> [6] "betweenss"    "size"         "iter"         "ifault"